+++ include/Collisions.h
@@ -8,10 +8,10 @@ namespace pheng {
 
     class Collisions {
     public:
-        //static double r_c;    =1? elastic collision, =0? perfect inelastic collision
+        //static float r_c;    =1? elastic collision, =0? perfect inelastic collision
 
         static bool objToObj(Object* o1, Object* o2, float r_c);
-        static bool circleToCircle(Circle* s1, Circle *s2, double restitution_coef);
+        static bool circleToCircle(Circle* s1, Circle *s2, float restitution_coef);
         static bool circleToRectangle(Circle *c, Rectangle *r);
 
         static bool areColliding(Object* o1, Object* o2);
