+++ include/Object.h
@@ -13,36 +13,36 @@ namespace pheng {
         vector2 Position;
         vector2 Velocity;
         vector2 Force = {0, 0};
-        double Mass;
+        float Mass;
 
-        double Energy;
+        float Energy;
 
-        double coef_st_frict = 0;
-        double coef_dn_frict = 0;
+        float coef_st_frict = 0;
+        float coef_dn_frict = 0;
 
-        double angular_velocity = 0;
+        float angular_velocity = 0;
 
         Mobility mobility;
 
         Object();
-        Object(vector2 pos, vector2 v, vector2 a, vector2 f, double m, Mobility mobility);
-        Object(vector2 pos, vector2 v, vector2 a, vector2 f, double m, double st_frict, double dn_frict);
+        Object(vector2 pos, vector2 v, vector2 a, vector2 f, float m, Mobility mobility);
+        Object(vector2 pos, vector2 v, vector2 a, vector2 f, float m, float st_frict, float dn_frict);
         virtual ~Object();
 
         void applyForce(vector2 force);
-        virtual void constraintsCollision(double constraints[], float r_f);
+        virtual void constraintsCollision(float constraints[], float r_f);
 
         // Gets overridden by each subclass
         virtual sf::Drawable& getDrawable();
 
-        virtual double getSize() = 0;
+        virtual float getSize() = 0;
         virtual vector2 getCenter() = 0;
 
         // Applies change to graphics
         virtual void applyChange();
 
         //Setters
-        void setMass(double kg);
+        void setMass(float kg);
         void setVelocity(vector2);
         void setPosition(vector2);
         void setMobility(Mobility);
@@ -52,10 +52,10 @@ namespace pheng {
         //Verlet integration
         vector2 Old_Position;
         vector2 Acceleration;
-        virtual void updatePositionVerlet(double dt) = 0;
+        virtual void updatePositionVerlet(float dt) = 0;
         virtual void accelerate(vector2 acceleration) = 0;
 
-        virtual float calculateEnergy(double h, double g);
+        virtual float calculateEnergy(float h, float g);
     };
 }
